+++ lib/src/jsMain/kotlin/platform.kt
@@ -2,6 +2,6 @@ package com.felipecsl.knes
 
 import kotlin.js.Date
 
-actual fun currentTimeMs(): Long {
-  return Date.now().toLong()
+actual fun currentTimeMs(): Double {
+  return Date.now()
 }
